requirements.txt
fastapi
uvicorn
youtube-transcript-api
python-dotenv

